daily

+++ leetcode/n0377-Combination_Sum_IV.cpp
@@ -6,7 +6,7 @@ public:
     int combinationSum4(vector<int>& nums, int target) {
 
         int n = nums.size();
-        vector<int> f(target+1);
+        vector<unsigned long long> f(target+1);
         f[0] = 1;
         for(int i = 1; i <= target; i++){
             for(int j = 0; j < n; j++){
